+++ App.js
@@ -1,6 +1,6 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import Login from './components/Login';
+import Login from './src/components/Login/Login';
 
 export default function App() {
   return (
