+++ src/components/Login/Login.styles.jsx
@@ -1,0 +1,24 @@
+import { StyleSheet } from "react-native"
+const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: '#fff',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    title:{
+        top:'100%'
+    },
+    textbox:{
+        width:'90%',
+        marginTop:'3%'
+    },
+    button:{
+        marginTop: '10%',
+        width:'85%'
+    },
+    text:{
+        marginTop:'5%',
+        textAlign:'left'
+    }
+  })
